requirements.txt
dash
plotly
pandas
sqlalchemy
python-dotenv
psycopg2-binary
flask
requests
